refactor: TokenType 경로 수정
